Transfer print logic to ObjectPrinter
